requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
requests
